feat(app-preview): fix router
