+++ packages/toolbars/preview/vite.config.ts
@@ -30,6 +30,9 @@ export default defineConfig({
       formats: ['es']
     },
     rollupOptions: {
+      output: {
+        banner: 'import "./style.css"'
+      },
       external: ['vue', /@opentiny\/tiny-engine.*/, /@opentiny\/vue.*/]
     }
   }
